# quizverse/backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "QuizVerse API",
        "description": "API dokumentacji dla aplikacji QuizVerse",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/quizzes/{quizId}/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Lista wszystkich komentarzy do quizu",
                "description": "Pobiera listę wszystkich komentarzy dla danego quizu, posortowanych od najnowszych",
                "operationId": "948771c254662d4c5f5283a335266ad4",
                "parameters": [
                    {
                        "name": "quizId",
                        "in": "path",
                        "description": "ID quizu",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista komentarzy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "quiz_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 5,
                                                "nullable": true
                                            },
                                            "user_name": {
                                                "type": "string",
                                                "example": "Jan Kowalski"
                                            },
                                            "content": {
                                                "type": "string",
                                                "example": "Świetny quiz!"
                                            },
                                            "rating": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Quiz nie znaleziony"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Utwórz nowy komentarz",
                "description": "Tworzy nowy komentarz do quizu. Uwierzytelniony użytkownik będzie przypisany do komentarza.",
                "operationId": "d2612b6c06f74546dbc1a005ecdf9a0e",
                "parameters": [
                    {
                        "name": "quizId",
                        "in": "path",
                        "description": "ID quizu",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content",
                                    "rating"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "maxLength": 1000,
                                        "minLength": 3,
                                        "example": "Świetny quiz, polecam!"
                                    },
                                    "rating": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1,
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Komentarz został utworzony",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "quiz_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 5,
                                            "nullable": true
                                        },
                                        "user_name": {
                                            "type": "string",
                                            "example": "Jan Kowalski"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "Świetny quiz, polecam!"
                                        },
                                        "rating": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Walidacja nie powiodła się"
                    },
                    "404": {
                        "description": "Quiz nie znaleziony"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/quizzes/{quizId}/comments/rating": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Średnia ocena quizu",
                "description": "Pobiera średnią ocenę i liczbę komentarzy dla danego quizu",
                "operationId": "f1457a9186a42f9cbabe9ed74e44b000",
                "parameters": [
                    {
                        "name": "quizId",
                        "in": "path",
                        "description": "ID quizu",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Średnia ocena i liczba komentarzy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "quiz_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "average_rating": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 4.5
                                        },
                                        "total_comments": {
                                            "type": "integer",
                                            "example": 12
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Quiz nie znaleziony"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}